// apps/web/src/shared/api/generated/model/loansControllerListType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type LoansControllerListType =
  (typeof LoansControllerListType)[keyof typeof LoansControllerListType];

export const LoansControllerListType = {
  borrowed_from_someone: "borrowed_from_someone",
  lent_to_someone: "lent_to_someone",
} as const;